move `build_archive_file_suggestion` to `extension.rs`
